Idemp, Review changes and Multi currency

+++ src/main/java/com/chipIn/ChipIn/dto/GroupDataByUserResponse.java
@@ -3,14 +3,13 @@ package com.chipIn.ChipIn.dto;
 import com.chipIn.ChipIn.entities.Group;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.UUID;
 
 @Data
 public class GroupDataByUserResponse {
-
     private Group group;
-    private Double amountOwedByUser;
+    /** In the group's default currency. */
+    private BigDecimal amountOwedByUser;
     private LocalDateTime lastExpenseDate;
-
 }
